Chop into equal-width intervals

chop_width() chops x into intervals of width width. chop_evenly chops x into intervals intervals of equal width.

chop_width(x, width, start, ..., left = width > 0)

chop_evenly(x, intervals, ..., groups, close_end = TRUE)

brk_width(width, start)

# S3 method for default
brk_width(width, start)

brk_evenly(intervals)

Arguments

x
A vector.

width
Width of intervals.

start
Leftpoint of first interval. By default the smallest finite x, or if width is negative, the largest finite x.

...
Passed to chop().

left
Passed to chop().

intervals
Integer: number of intervals to create.

groups
Do not use.

close_end
Passed to chop().

Value

For chop_* functions, a factor of the same length as x.

Details

If width is negative, intervals will go downwards from start.

chop_evenly sets close_end = TRUE by default. chop_width sets left = FALSE if width is negative.

See also

Examples

chop_width(1:10, 2)
#>  [1] [1, 3)  [1, 3)  [3, 5)  [3, 5)  [5, 7)  [5, 7)  [7, 9)  [7, 9)  [9, 11)
#> [10] [9, 11)
#> Levels: [1, 3) [3, 5) [5, 7) [7, 9) [9, 11)

chop_width(1:10, 2, start = 0)
#>  [1] [0, 2)  [2, 4)  [2, 4)  [4, 6)  [4, 6)  [6, 8)  [6, 8)  [8, 10) [8, 10)
#> [10] {10}   
#> Levels: [0, 2) [2, 4) [4, 6) [6, 8) [8, 10) {10}

chop_width(1:9, -2)
#> [1] {1}    (1, 3] (1, 3] (3, 5] (3, 5] (5, 7] (5, 7] (7, 9] (7, 9]
#> Levels: {1} (1, 3] (3, 5] (5, 7] (7, 9]

chop(1:10, brk_width(2, 0))
#>  [1] [0, 2)  [2, 4)  [2, 4)  [4, 6)  [4, 6)  [6, 8)  [6, 8)  [8, 10) [8, 10)
#> [10] {10}   
#> Levels: [0, 2) [2, 4) [4, 6) [6, 8) [8, 10) {10}

chop_evenly(0:10, 5)
#>  [1] [0, 2)  [0, 2)  [2, 4)  [2, 4)  [4, 6)  [4, 6)  [6, 8)  [6, 8)  [8, 10]
#> [10] [8, 10] [8, 10]
#> Levels: [0, 2) [2, 4) [4, 6) [6, 8) [8, 10]