Label chopped intervals like 1 - 3, 4 - 5, ...

This label style is user-friendly, but doesn't distinguish between left- and right-closed intervals.

lbl_dash(symbol = " - ", raw = FALSE, fmt = NULL)

Arguments

symbol
String: symbol to use for the dash.

raw
Logical. Always use raw breaks in labels, rather than e.g. quantiles or standard deviations?

fmt
A format. Can be a string, passed into base::sprintf() or format() methods; or a one-argument formatting function.

Value

A vector of labels for chop, or a function that creates labels.

See also

Examples

chop(1:10, c(2, 5, 8), lbl_dash())

#>  [1] 1 - 2  2 - 5  2 - 5  2 - 5  5 - 8  5 - 8  5 - 8  8 - 10 8 - 10 8 - 10
#> Levels: 1 - 2 2 - 5 5 - 8 8 - 10

chop(1:10, c(2, 5, 8), lbl_dash(" to ", fmt = "%.1f"))

#>  [1] 1.0 to 2.0  2.0 to 5.0  2.0 to 5.0  2.0 to 5.0  5.0 to 8.0  5.0 to 8.0 
#>  [7] 5.0 to 8.0  8.0 to 10.0 8.0 to 10.0 8.0 to 10.0
#> Levels: 1.0 to 2.0 2.0 to 5.0 5.0 to 8.0 8.0 to 10.0

pretty <- function (x) prettyNum(x, big.mark = ",", digits = 1)
chop(runif(10) * 10000, c(3000, 7000), lbl_dash(" to ", fmt = pretty))

#>  [1] 74 to 3,000    7,000 to 8,343 3,000 to 7,000 74 to 3,000    74 to 3,000   
#>  [6] 3,000 to 7,000 3,000 to 7,000 74 to 3,000    7,000 to 8,343 7,000 to 8,343
#> Levels: 74 to 3,000 3,000 to 7,000 7,000 to 8,343