Label chopped intervals by their left or right endpoints

This is useful when the left endpoint unambiguously indicates the interval.

lbl_endpoint(fmt = NULL, raw = FALSE, left = TRUE)

Arguments

fmt
A format. Can be a string, passed into base::sprintf() or format() methods; or a one-argument formatting function.

raw
Logical. Always use raw breaks in labels, rather than e.g. quantiles or standard deviations?

left
Flag. Use left endpoint or right endpoint?

Value

A vector of labels for chop, or a function that creates labels.

Examples

chop(1:10, c(2, 5, 8), lbl_endpoint(left = TRUE))

#>  [1] 1 2 2 2 5 5 5 8 8 8
#> Levels: 1 2 5 8

chop(1:10, c(2, 5, 8), lbl_endpoint(left = FALSE))

#>  [1] 2  5  5  5  8  8  8  10 10 10
#> Levels: 2 5 8 10

if (requireNamespace("lubridate")) {
  tab_width(
          as.Date("2000-01-01") + 0:365,
         months(1),
         labels = lbl_endpoint(fmt = "%b")
       )
}

#> x
#> Jan Feb Mar Apr May Jun Jul Aug Sep Oct Nov Dec 
#>  31  29  31  30  31  30  31  31  30  31  30  31