Cut numeric data into intervals

chop cuts x into intervals. It returns a factor of the same length as x, representing which interval contains each element of x.

chop(
  x,
  breaks,
  labels,
  extend = NULL,
  left = TRUE,
  close_end = FALSE,
  drop = TRUE
)

kiru(
  x,
  breaks,
  labels,
  extend = NULL,
  left = TRUE,
  close_end = FALSE,
  drop = TRUE
)

Arguments

x
A numeric vector.

breaks
See below.

labels
See below.

extend
Logical. Extend breaks to +/-Inf?

left
Logical. Left-closed breaks?

close_end
Logical. Close last break at right? (If left is FALSE, close first break at left?)

drop
Logical. Drop unused levels from the result?

Value

A factor of the same length as x, representing the intervals containing the value of x.

Details

breaks may be a numeric vector or a function.

If it is a vector, breaks gives the break endpoints. Repeated values create singleton intervals. For example breaks = c(1, 3, 3, 5) creates 3 intervals: [1, 3), {3} and (3, 5].

By default, left-closed intervals are created. If left is FALSE, right- closed intervals are created.

If close_end is TRUE the end break will be closed at both ends, ensuring that all values y with min(x) <= y <= max(x) are included in the default intervals. That is:

If left is TRUE and close_end is TRUE, breaks will look like [x1, x2), [x2, x3) ... [x_n-1, x_n].

If left is FALSE and close_end is TRUE, breaks will look like [x1, x2], (x2, x3] ... (x_n-1, x_n].

If left is TRUE and close_end is FALSE, all breaks will look like ...[x1, x2) ...

If left is FALSE and close_end is FALSE, all breaks will look like ...(x1, x2] ...

If breaks is a function it is called with the x, extend, left and close_end arguments, and should return an object of class breaks. Use brk_ functions in this context, to create a variety of data-dependent breaks.

labels may be a character vector. It should have the same length as the number of intervals. Alternatively, use a lbl_ function such as [lbl_seq()].

If extend is TRUE, intervals will be extended to [-Inf, min(breaks)) and (max(breaks), Inf].

If extend is NULL (the default), intervals will be extended to [min(x), min(breaks)) and (max(breaks), max(x)], only if necessary -- i.e. if min(x) < min(breaks) and max(x) > max(breaks) respectively.

Extending intervals, either by extend = NULL or extend = FALSE, always leaves the central, non-extended intervals unchanged. In particular, close_end applies to the central intervals, not to the extended ones. For example, if breaks = c(1, 3, 5) and close_end = TRUE, the resulting breaks will be

[1, 3), [3, 5]

and if they are extended on both ends the result will be e.g.

[-Inf, 1), [1, 3), [3, 5], (5, Inf]

NA values in x, and values which are outside the (extended) endpoints, return NA.

Note that chop, like all of R, uses binary arithmetic. Thus, numbers may not be exactly equal to what you think they should be. There is an example below.

[x1, x2) ...`

If left is FALSE and close_end is FALSE, all breaks will look like `...(x1, x2]: R:x1,%20x2)%20...%60%0A*%20If%20%60left%60%20is%20%60FALSE%60%20and%20%60close_end%60%20is%20%60FALSE%60,%20all%20breaks%20will%20look%20like%0A%20%20%20%60...(x1,%20x2 [lbl_seq()]: R:lbl_seq() [3, 5]: R:3,%205 [3, 5]: R:3,%205

kiru is a synonym for chop. If you load tidyr, you can use it to avoid confusion with tidyr::chop().

See also

Examples

chop(1:3, 2)

#> [1] [1, 2) [2, 3] [2, 3]
#> Levels: [1, 2) [2, 3]

chop(1:10, c(2, 5, 8))

#>  [1] [1, 2)  [2, 5)  [2, 5)  [2, 5)  [5, 8)  [5, 8)  [5, 8)  [8, 10] [8, 10]
#> [10] [8, 10]
#> Levels: [1, 2) [2, 5) [5, 8) [8, 10]

chop(1:10, c(2, 5, 8), extend = FALSE)

#>  [1] <NA>   [2, 5) [2, 5) [2, 5) [5, 8) [5, 8) [5, 8) <NA>   <NA>   <NA>  
#> Levels: [2, 5) [5, 8)

chop(1:10, c(2, 5, 5, 8))

#>  [1] [1, 2)  [2, 5)  [2, 5)  [2, 5)  {5}     (5, 8)  (5, 8)  [8, 10] [8, 10]
#> [10] [8, 10]
#> Levels: [1, 2) [2, 5) {5} (5, 8) [8, 10]

chop(1:10, c(2, 5, 8), left = FALSE)

#>  [1] [1, 2]  [1, 2]  (2, 5]  (2, 5]  (2, 5]  (5, 8]  (5, 8]  (5, 8]  (8, 10]
#> [10] (8, 10]
#> Levels: [1, 2] (2, 5] (5, 8] (8, 10]

chop(1:10, c(2, 5, 8), close_end = TRUE)

#>  [1] [1, 2)  [2, 5)  [2, 5)  [2, 5)  [5, 8]  [5, 8]  [5, 8]  [5, 8]  (8, 10]
#> [10] (8, 10]
#> Levels: [1, 2) [2, 5) [5, 8] (8, 10]

chop(1:10, brk_quantiles(c(0.25, 0.75)))

#>  [1] [0%, 25%)   [0%, 25%)   [0%, 25%)   [25%, 75%)  [25%, 75%)  [25%, 75%) 
#>  [7] [25%, 75%)  [75%, 100%] [75%, 100%] [75%, 100%]
#> Levels: [0%, 25%) [25%, 75%) [75%, 100%]

chop(1:10, c(2, 5, 8), labels = lbl_dash())

#>  [1] 1 - 2  2 - 5  2 - 5  2 - 5  5 - 8  5 - 8  5 - 8  8 - 10 8 - 10 8 - 10
#> Levels: 1 - 2 2 - 5 5 - 8 8 - 10

# floating point inaccuracy:
chop(0.3/3, c(0, 0.1, 0.1, 1))

#> [1] [0, 0.1)
#> Levels: [0, 0.1)