Cut numeric data into intervals

chop cuts x into intervals. It returns a factor of the same length as x, representing which interval contains each element of x.

chop(x, breaks, labels, extend = NULL, drop = TRUE)

kiru(x, breaks, labels, extend = NULL, drop = TRUE)

Arguments

x
A numeric vector.

breaks
See below.

labels
See below.

extend
Logical. Extend breaks to +/-Inf?

drop
Logical. Drop unused levels from the result?

Value

A factor of the same length as x, representing the intervals containing the value of x.

Details

breaks may be a numeric vector or a function.

If it is a vector, breaks gives the break endpoints.Repeated values create singleton intervals. For example breaks = c(1, 3, 3, 5) creates 3 intervals: [1, 3), {3} and (3, 5]. Default breaks are left-closed except for the last interval; for right-closed breaks see brk_right().

If breaks is a function it is called with a single argument, x, and returns an object of class breaks.

labels may be a character vector. It should have the same length as the number of intervals. Alternatively, use a lbl_ function such as lbl_numerals().

If extend is TRUE, intervals will be extended to [-Inf, min(breaks)) and (max(breaks), Inf].

If extend is NULL (the default), intervals will be extended to [min(x), min(breaks)) and (max(breaks), max(x)], only if necessary -- i.e. if min(x) < min(breaks) and max(x) > max(breaks) respectively.

NA values in x, and values which are outside the (extended) endpoints, return NA.

Note that chop, like all of R, uses binary arithmetic. Thus, numbers may not be exactly equal to what you think they should be. There is an example below.

kiru is a synonym for chop. If you load tidyr, you can use it to avoid confusion with tidyr::chop().

See also

Examples

chop(1:3, 2)

#> [1] [1, 2] [1, 2] (2, 3]
#> Levels: [1, 2] (2, 3]

chop(1:10, c(2, 5, 8))

#>  [1] [1, 2)  [2, 5)  [2, 5)  [2, 5)  [5, 8]  [5, 8]  [5, 8]  [5, 8]  (8, 10]
#> [10] (8, 10]
#> Levels: [1, 2) [2, 5) [5, 8] (8, 10]

chop(1:10, c(2, 5, 8), extend = FALSE)

#>  [1] <NA>   [2, 5) [2, 5) [2, 5) [5, 8] [5, 8] [5, 8] [5, 8] <NA>   <NA>  
#> Levels: [2, 5) [5, 8]

chop(1:10, c(2, 5, 5, 8))

#>  [1] [1, 2)  [2, 5)  [2, 5)  [2, 5)  {5}     (5, 8]  (5, 8]  (5, 8]  (8, 10]
#> [10] (8, 10]
#> Levels: [1, 2) [2, 5) {5} (5, 8] (8, 10]

chop(1:10, brk_quantiles(c(0.25, 0.75)))

#>  [1] [0%, 25%)   [0%, 25%)   [0%, 25%)   [25%, 75%]  [25%, 75%]  [25%, 75%] 
#>  [7] [25%, 75%]  (75%, 100%] (75%, 100%] (75%, 100%]
#> Levels: [0%, 25%) [25%, 75%] (75%, 100%]

chop(1:10, c(2, 5, 8), labels = lbl_dash())

#>  [1] 1 - 2  2 - 5  2 - 5  2 - 5  5 - 8  5 - 8  5 - 8  5 - 8  8 - 10 8 - 10
#> Levels: 1 - 2 2 - 5 5 - 8 8 - 10

# floating point inaccuracy:
chop(0.3/3, c(0, 0.1, 0.1, 1))

#> [1] [0, 0.1)
#> Levels: [0, 0.1)