Cut numeric data into intervals

chop cuts x into intervals. It returns a factor of the same length as x, representing which interval contains each element of x.

chop(x, breaks, labels = lbl_intervals(), extend = TRUE, drop = TRUE)

Arguments

x
A numeric vector.

breaks
See below.

labels
See below.

extend
Logical. Extend breaks to +/-Inf?

drop
Logical. Drop unused levels from the result?

Value

A factor, or an integer if labels is NULL, of the same length as x, representing the intervals containing the value of x.

Details

breaks may be a numeric vector, an object of class breaks, or a function.

If it is a vector, breaks gives the break endpoints.Repeated values create singleton intervals. For example breaks = c(1, 3, 3, 5) creates 3 intervals: [1, 3), {3} and (3, 5]. Default breaks are left-closed except the end; for right-closed breaks see brk_right().

If breaks is a function it is called with a single argument, x, and returns an object of class breaks.

labels may be a character vector or a function. If it is a character vector it should have the same length as the number of intervals - i.e. one more than length(breaks) if extend is TRUE, one less otherwise.

If labels is a function it should take an object of class breaks and return an appropriate character vector.

If extend is TRUE, intervals will be extended to [-Inf, min(breaks)) and (max(breaks), Inf], unless those endpoints are already infinite.

NA values in x, and values which are outside the (extendeD) endpoints, return NA.

See also

Examples

chop(1:3, 2)

#> [1] [-Inf, 2] [-Inf, 2] (2, Inf] 
#> Levels: [-Inf, 2] (2, Inf]

chop(rnorm(10), -2:2)

#>  [1] [-1, 0)  [0, 1)   (2, Inf] [-2, -1) [0, 1)   [-2, -1) [-1, 0)  [-1, 0) 
#>  [9] [0, 1)   [-1, 0) 
#> Levels: [-2, -1) [-1, 0) [0, 1) (2, Inf]

chop(rnorm(10), -2:2, extend = FALSE)

#>  [1] [0, 1)   [0, 1)   [-2, -1) [0, 1)   [1, 2]   [-1, 0)  [-1, 0)  [-1, 0) 
#>  [9] [-1, 0)  [-2, -1)
#> Levels: [-2, -1) [-1, 0) [0, 1) [1, 2]

chop(rnorm(10), brk_quantiles(c(0.25, 0.75)))

#>  [1] (0.76, Inf]   [-0.24, 0.76] [-0.24, 0.76] (0.76, Inf]   [-0.24, 0.76]
#>  [6] [-0.24, 0.76] [-Inf, -0.24) [-Inf, -0.24) [-Inf, -0.24) (0.76, Inf]  
#> Levels: [-Inf, -0.24) [-0.24, 0.76] (0.76, Inf]

chop(rnorm(10), -2:2, labels = lbl_dash())

#>  [1] 0 - 1   -1 - 0  -1 - 0  -1 - 0  0 - 1   2 - Inf 0 - 1   0 - 1   0 - 1  
#> [10] -2 - -1
#> Levels: -2 - -1 -1 - 0 0 - 1 2 - Inf