Cut data into intervals

chop cuts x into intervals. It returns a factor of the same length as x, representing which interval contains each element of x.

Usage

chop(
  x,
  breaks,
  labels = lbl_intervals(),
  extend = NULL,
  left = TRUE,
  close_end = FALSE,
  drop = TRUE
)

kiru(
  x,
  breaks,
  labels = lbl_intervals(),
  extend = NULL,
  left = TRUE,
  close_end = FALSE,
  drop = TRUE
)

Arguments

x

A vector.

breaks

See below.

labels

See below.

extend

Logical. Extend breaks to +/-Inf?

left

Logical. Left-closed breaks?

close_end

Logical. Close last break at right? (If left is FALSE, close first break at left?)

drop

Logical. Drop unused levels from the result?

Value

A factor of the same length as x, representing the intervals containing the value of x.

Details

x may be numeric, or a Date or Date-Time.

Breaks

breaks may be a numeric vector or a function.

If it is a vector, breaks gives the break endpoints. Repeated values create singleton intervals. For example breaks = c(1, 3, 3, 5) creates 3 intervals: [1, 3), {3} and (3, 5].

If breaks is a function it is called with the x, extend, left and close_end arguments, and should return an object of class breaks. Use brk_ functions in this context, to create a variety of data-dependent breaks.

Options for breaks

By default, left-closed intervals are created. If left is FALSE, right- closed intervals are created.

If close_end is TRUE the end break will be closed at both ends, ensuring that all values y with min(x) <= y <= max(x) are included in the default intervals.

Overall:

If left is TRUE and close_end is TRUE, breaks will look like [x1, x2), [x2, x3) ... [x_n-1, x_n].

If left is FALSE and close_end is TRUE, breaks will look like [x1, x2], (x2, x3] ... (x_n-1, x_n].

If left is TRUE and close_end is FALSE, all breaks will look like ...[x1, x2) ....

If left is FALSE and close_end is FALSE, all breaks will look like ...(x1, x2] ....

Extending intervals

If extend is TRUE, intervals will be extended to [-Inf, min(breaks)) and (max(breaks), Inf].

If extend is NULL (the default), intervals will be extended to [min(x), min(breaks)) and (max(breaks), max(x)], only if necessary -- i.e. if min(x) < min(breaks) and max(x) > max(breaks) respectively.

Extending intervals, either by extend = NULL or extend = FALSE, always leaves the central, non-extended intervals unchanged. In particular, close_end applies to the central intervals, not to the extended ones. For example, if breaks = c(1, 3, 5) and close_end = TRUE, the resulting breaks will be

[1, 3), [3, 5]

and if extend = TRUE the result will be

[-Inf, 1), [1, 3), [3, 5], (5, Inf]

Labels

labels may be a character vector. It should have the same length as the number of intervals. Alternatively, use a lbl_ function such as lbl_seq().

If labels is NULL, then integer codes will be returned instead of a factor.

Miscellaneous

NA values in x, and values which are outside the extended endpoints, return NA.

kiru is a synonym for chop. If you load tidyr, you can use it to avoid confusion with tidyr::chop().

Note that chop, like all of R, uses binary arithmetic. Thus, numbers may not be exactly equal to what you think they should be. There is an example below.

See also

Examples

chop(1:3, 2)
#> [1] [1, 2) [2, 3] [2, 3]
#> Levels: [1, 2) [2, 3]

chop(1:10, c(2, 5, 8))
#>  [1] [1, 2)  [2, 5)  [2, 5)  [2, 5)  [5, 8)  [5, 8)  [5, 8)  [8, 10] [8, 10]
#> [10] [8, 10]
#> Levels: [1, 2) [2, 5) [5, 8) [8, 10]

chop(1:10, c(2, 5, 8), extend = FALSE)
#>  [1] <NA>   [2, 5) [2, 5) [2, 5) [5, 8) [5, 8) [5, 8) <NA>   <NA>   <NA>  
#> Levels: [2, 5) [5, 8)

chop(1:10, c(2, 5, 5, 8))
#>  [1] [1, 2)  [2, 5)  [2, 5)  [2, 5)  {5}     (5, 8)  (5, 8)  [8, 10] [8, 10]
#> [10] [8, 10]
#> Levels: [1, 2) [2, 5) {5} (5, 8) [8, 10]

chop(1:10, c(2, 5, 8), left = FALSE)
#>  [1] [1, 2]  [1, 2]  (2, 5]  (2, 5]  (2, 5]  (5, 8]  (5, 8]  (5, 8]  (8, 10]
#> [10] (8, 10]
#> Levels: [1, 2] (2, 5] (5, 8] (8, 10]

chop(1:10, c(2, 5, 8), close_end = TRUE)
#>  [1] [1, 2)  [2, 5)  [2, 5)  [2, 5)  [5, 8]  [5, 8]  [5, 8]  [5, 8]  (8, 10]
#> [10] (8, 10]
#> Levels: [1, 2) [2, 5) [5, 8] (8, 10]

chop(1:10, brk_quantiles(c(0.25, 0.75)))
#>  [1] [0%, 25%)   [0%, 25%)   [0%, 25%)   [25%, 75%)  [25%, 75%)  [25%, 75%) 
#>  [7] [25%, 75%)  [75%, 100%] [75%, 100%] [75%, 100%]
#> Levels: [0%, 25%) [25%, 75%) [75%, 100%]

chop(1:10, c(2, 5, 8), labels = lbl_dash())
#>  [1] 1—2  2—5  2—5  2—5  5—8  5—8  5—8  8—10 8—10 8—10
#> Levels: 1—2 2—5 5—8 8—10

# floating point inaccuracy:
chop(0.3/3, c(0, 0.1, 0.1, 1), labels = c("< 0.1", "0.1", "> 0.1"))
#> [1] < 0.1
#> Levels: < 0.1