Chop by quantiles

chop_quantiles chops data by quantiles. chop_equally chops data into equal-sized groups. chop_deciles is a convenience shortcut and chops into deciles.

chop_quantiles(x, probs, ..., left = is.numeric(x), close_end = TRUE)

chop_deciles(x, ...)

chop_equally(x, groups, ..., left = is.numeric(x), close_end = TRUE)

brk_quantiles(probs, ...)

brk_equally(groups)

Arguments

x
A vector.

probs
A vector of probabilities for the quantiles.

...
Passed to chop(), or for brk_quantiles to stats::quantile().

left
Passed to chop().

close_end
Passed to chop().

groups
Number of groups.

Value

For chop_* functions, a factor of the same length as x.

Details

Note that these functions set close_end = TRUE by default. This helps ensure that e.g. chop_quantiles(x, c(0, 1/3, 2/3, 1) will split the data into three equal-sized groups.

For non-numeric x, left is set to FALSE by default. This works better for calculating "type 1" quantiles, since they round down. See stats::quantile().

See also

Examples

chop_quantiles(1:10, 1:3/4)
#>  [1] [0%, 25%)   [0%, 25%)   [0%, 25%)   [25%, 50%)  [25%, 50%)  [50%, 75%] 
#>  [7] [50%, 75%]  (75%, 100%] (75%, 100%] (75%, 100%]
#> Levels: [0%, 25%) [25%, 50%) [50%, 75%] (75%, 100%]

chop(1:10, brk_quantiles(1:3/4))
#>  [1] [0%, 25%)   [0%, 25%)   [0%, 25%)   [25%, 50%)  [25%, 50%)  [50%, 75%) 
#>  [7] [50%, 75%)  [75%, 100%] [75%, 100%] [75%, 100%]
#> Levels: [0%, 25%) [25%, 50%) [50%, 75%) [75%, 100%]

chop_deciles(1:10)
#>  [1] [0%, 10%)   [10%, 20%)  [20%, 30%)  [30%, 40%)  [40%, 50%)  [50%, 60%) 
#>  [7] [60%, 70%)  [70%, 80%)  [80%, 90%)  [90%, 100%]
#> 10 Levels: [0%, 10%) [10%, 20%) [20%, 30%) [30%, 40%) [40%, 50%) ... [90%, 100%]

chop_equally(1:10, 5)
#>  [1] [0%, 20%)   [0%, 20%)   [20%, 40%)  [20%, 40%)  [40%, 60%)  [40%, 60%) 
#>  [7] [60%, 80%)  [60%, 80%)  [80%, 100%] [80%, 100%]
#> Levels: [0%, 20%) [20%, 40%) [40%, 60%) [60%, 80%) [80%, 100%]

# to label by the quantiles themselves:
chop_quantiles(1:10, 1:3/4, lbl_intervals(raw = TRUE))
#>  [1] [1, 3.25)   [1, 3.25)   [1, 3.25)   [3.25, 5.5) [3.25, 5.5) [5.5, 7.75]
#>  [7] [5.5, 7.75] (7.75, 10]  (7.75, 10]  (7.75, 10] 
#> Levels: [1, 3.25) [3.25, 5.5) [5.5, 7.75] (7.75, 10]